Fix Error: uv_signal_start EINVAL

+++ src/PLC.ts
@@ -232,9 +232,11 @@ class PLC extends EIPSocketPool implements IDeviceProps {
       const onExit = () => {
         this.close();
       };
-      process.once("SIGINT", onExit); // Ctrl + C
-      process.once("SIGKILL", onExit); // kill process
       process.once("beforeExit", onExit); // on process.exit
+      process.once("SIGINT", onExit); // Ctrl + C
+      if (parseInt(process.versions.node) >= 10) {
+        process.once("SIGKILL", onExit); // kill process
+      }
     }
   }
   /**
